requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pybit>=5.6.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
ta>=0.11.0
colorlog>=6.7.0
httpx>=0.25.0
python-multipart>=0.0.6
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
feedparser>=6.0.10
textblob>=0.17.1
pymongo
dnspython>=2.4.0
discord.py>=2.3.0
